tests: use custom test-driver for specific tests
This patch is just used to showcase what can be done
with a custom test-driver, how it could improve testing
and logging of information, and maybe allow for
benchmarking of each test.

Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/10951
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>

+++ tests/test_launch_daemon.sh
@@ -10,11 +10,17 @@ function invoke_daemon()
     $LOG_COMPILER $LOG_FLAGS $phobosd "$*" &
     wait $! || { echo "failed to start daemon"; return 1; }
     PID_DAEMON=`cat $PHOBOSD_PID_FILEPATH`
-    rm $PHOBOSD_PID_FILEPATH
 }
 
 function waive_daemon()
 {
+    if [ ! -f "$PHOBOSD_PID_FILEPATH" ]; then
+        return 0
+    fi
+
+    PID_DAEMON=`cat $PHOBOSD_PID_FILEPATH`
+    rm $PHOBOSD_PID_FILEPATH
+
     if [ $PID_DAEMON -eq 0 ]; then
         return 0
     fi
